feat(ai_reviewer): 重构AI审查器为模块化架构
- 新增 ai_reviewer 模块初始化文件，保持向后兼容的导入方式
- 拆分原单文件为专门模块：api_client、batch_processor、compression、constants、label_recommender、prompt_builder、result_parser、reviewer、tools
- 迁移 API 客户端逻辑至 api_client.py，封装带重试机制的 API 调用
- 迁移批处理逻辑至 batch_processor.py，包含文件分批、审查批处理、结果归约等功能
- 新增常量定义模块 constants.py，统一管理配置参数
- 导出所有子模块和主类，支持细粒度控制和向后兼容
