+++ backend/services/ai_reviewer/compression/__init__.py
@@ -1,0 +1,8 @@
+"""上下文压缩模块
+
+导出压缩相关的类。
+"""
+
+from .context_compressor import ContextCompressor
+
+__all__ = ["ContextCompressor"]
