+++ backend/services/ai_reviewer/tools/__init__.py
@@ -1,0 +1,16 @@
+"""工具模块
+
+导出工具相关的类。
+"""
+
+from .file_tool import FileToolHandler
+from .handler import ToolHandler
+from .manager import ToolManager
+from .search_tool import SearchToolHandler
+
+__all__ = [
+    "FileToolHandler",
+    "SearchToolHandler",
+    "ToolHandler",
+    "ToolManager",
+]
